+++ debug/plugins/apexcharts/src/svgjs/svg.js
@@ -1507,6 +1507,8 @@
 
 
 
+
+
         // functionToCall: [list of morphable objects]
         // e.g. move: [SVG.Number, SVG.Number]
       };this.attrs = {
@@ -1636,6 +1638,7 @@
         if (element instanceof SVG.Element) {var box; // yes this is ugly, but Firefox can be a pain when it comes to elements that are not yet rendered
           try {if (!document.documentElement.contains) {// This is IE - it does not support contains() for top-level SVGs
               var topParent = element.node;while (topParent.parentNode) {topParent = topParent.parentNode;}if (topParent != document) throw new Error('Element not in the dom');} else {
+
 
 
 
@@ -2700,8 +2703,6 @@
 
 
 
-
-
       // Get all siblings, including myself
     });SVG.Gradient = SVG.invent({ // Initialize node
       create: function (type) {this.constructor.call(this, SVG.create(type + 'Gradient')); // store type
@@ -2774,8 +2775,7 @@
       cy: function (y) {return y == null ? this.attr('cy') : this.attr('cy', y);}, // Set width of element
       width: function (width) {return width == null ? this.rx() * 2 : this.rx(new SVG.Number(width).divide(2));}, // Set height of element
       height: function (height) {return height == null ? this.ry() * 2 : this.ry(new SVG.Number(height).divide(2));}, // Custom size function
-      size: function (width, height) {var p = proportionalSize(this, width, height);return this.rx(new SVG.Number(p.width).divide(2)).ry(new SVG.Number(p.height).divide(2));}
-    });
+      size: function (width, height) {var p = proportionalSize(this, width, height);return this.rx(new SVG.Number(p.width).divide(2)).ry(new SVG.Number(p.height).divide(2));} });
   SVG.Line = SVG.invent({
     // Initialize node
     create: 'line',