+++ debug/t1/manager/tpl/list.js
@@ -25,7 +25,7 @@ class Template {
                             <tr>
                                 <th>
                                     <label>
-                                        <input type="checkbox" class="checkbox" name="chkall" id="chkall">
+                                        <input type="checkbox" class="checkbox chkid" name="chkall" id="chkall">
                                     </label>
                                 </th>
                                 <th>성명</th>